Adicionada opção de buscar, editar e deletar 1 usuario

+++ localpe-backend/src/main/java/br/com/strawhats/localpe/controllers/LugarController.java
@@ -1,10 +1,11 @@
 package br.com.strawhats.localpe.controllers;
 
 import br.com.strawhats.localpe.models.Categoria;
+import br.com.strawhats.localpe.models.Comentario;
 import br.com.strawhats.localpe.models.Lugar;
+import br.com.strawhats.localpe.models.Usuario;
 import br.com.strawhats.localpe.services.LugarService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
@@ -44,7 +45,7 @@ public class LugarController {
     }
 
     @PutMapping("/lugar")
-    public void editarLugar(Lugar lugar){
+    public void editarLugar(@RequestHeader Lugar lugar){
         try {
             lugarService.cadastrarLugar(lugar);
         }catch (Exception e){
@@ -52,10 +53,15 @@ public class LugarController {
         }
 
     }
-    @DeleteMapping("/lugar")
-    public void deletarLugar(Long id){
+    @DeleteMapping("/lugar/{id}")
+    public void deletarLugar(@PathVariable Long id, @RequestHeader Usuario usuario){
         try {
+            if(usuario.getAdmin()){
             lugarService.excluirLugar(id);
+            }
+            else {
+                throw new Exception("Usuario não tem permisssão");
+            }
         }catch (Exception e){
             System.out.println("erro: "+ e.getMessage());
         }
@@ -66,4 +72,11 @@ public class LugarController {
     public List<Lugar> listarPorCategoria(@PathVariable Categoria categoria){
         return lugarService.listarPorCategoria(categoria);
     }
+
+    @GetMapping("/comentarios/{id}")
+    public List<Comentario> listarComentarios(@PathVariable Long LugarId){
+        return lugarService.listarComentarios(LugarId);
+    }
+
+
 }
